- Merge `customizable-e2e-test` test suite back together as per review comments. - Run tests sequentially to avoid concurrent writes for `customizable-e2e-test`.
