--- samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-aad/testConfig.json
@@ -1,1 +1,0 @@
-{"msalConfig":{"auth":{"clientId":"3fba556e-5d4a-48e3-8e1a-fd57c12cb82e","authority":"https://login.windows-ppe.net/common/"},"cache":{"cacheLocation":"sessionStorage","storeAuthStateInCookie":false}},"request":{"scopes":["User.Read"],"redirectStartPage":"/"}}
