--- samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-b2c/testConfig.json
@@ -1,1 +1,0 @@
-{"msalConfig":{"auth":{"clientId":"4c837770-7a2b-471e-aafa-3328d04a23b1","authority":"https://msidlabb2c.b2clogin.com/msidlabb2c.onmicrosoft.com/B2C_1_SISOPolicy/","knownAuthorities":["msidlabb2c.b2clogin.com"]},"cache":{"cacheLocation":"sessionStorage","storeAuthStateInCookie":false}},"request":{"scopes":["https://msidlabb2c.onmicrosoft.com/4c837770-7a2b-471e-aafa-3328d04a23b1/read"]}}
