--- samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-local-storage/index.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <title>Quickstart | MSAL.JS Vanilla JavaScript SPA</title>
-
-    <script src="../../lib/msal-browser.js"></script>
-
-    <!-- adding Bootstrap 4 for UI components  -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-    <link rel="SHORTCUT ICON" href="https://c.s-microsoft.com/favicon.ico?v2" type="image/x-icon">
-  </head>
-  <body>
-    <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-      <a class="navbar-brand" href="/">MS Identity Platform</a>
-      <div class="btn-group ml-auto dropleft">
-          <button type="button" id="SignIn" class="btn btn-secondary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-            Sign In
-          </button>
-          <div class="dropdown-menu">
-            <button class="dropdown-item" id="popup" onclick="signIn(this.id)">Sign in using Popup</button>
-            <button class="dropdown-item" id="redirect" onclick="signIn(this.id)">Sign in using Redirect</button>
-          </div>
-      </div>
-    </nav>
-    <br>
-    <h5 class="card-header text-center">Vanilla JavaScript SPA calling MS Graph API with MSAL.JS</h5>
-    <br>
-    <div class="row" style="margin:auto" >
-    <div id="card-div" class="col-md-3" style="display:none">
-    <div class="card text-center">
-      <div class="card-body">
-        <h5 class="card-title" id="WelcomeMessage">Please sign-in to see your profile and read your mails</h5>
-        <div id="profile-div"></div>
-        <br>
-        <br>
-        <button class="btn btn-primary" id="acquireTokenRedirect" onclick="getTokenRedirect()">acquireTokenRedirect</button>
-        <br>
-        <br>
-        <button class="btn btn-primary" id="acquireTokenPopup" onclick="getTokenPopup()">acquireTokenPopup</button>
-        <br>
-        <br>
-        <button class="btn btn-primary" id="acquireTokenSilent" onclick="getTokenSilently()">acquireTokenSilent</button>
-      </div>
-    </div>
-    </div>
-    <br>
-    <br>
-      <div class="col-md-4">
-        <div class="list-group" id="list-tab" role="tablist">
-        </div>
-      </div>
-      <div class="col-md-5">
-        <div class="tab-content" id="nav-tabContent">
-        </div>
-      </div>
-    </div>
-    <br>
-    <br>
-
-    <!-- importing bootstrap.js and supporting js libraries -->
-    <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
-
-    <!-- importing app scripts | load order is important -->
-    <script type="text/javascript" src="ui.js"></script>
-    <script type="text/javascript" src="auth.js"></script>
-  </body>
-</html>
