+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-test/auth.js
