+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-test/jest.config.js
@@ -1,8 +1,8 @@
 module.exports = {
-    displayName: "Vanilla JS customizable e2e aad",
+    displayName: "Vanilla JS customizable e2e test",
     globals: {
         __PORT__: 30661,
-        __STARTCMD__: "npm start -- --port 30661 --sample customizable-e2e-aad",
+        __STARTCMD__: "npm start -- --port 30661 --sample customizable-e2e-test",
     },
     testMatch: [
         "<rootDir>/test/**/*.spec.ts",
