+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-test/test/localStorage.spec.ts
@@ -4,12 +4,12 @@ import { BrowserCacheUtils } from "../../../../../e2eTestUtils/BrowserCacheTestU
 import { LabApiQueryParams } from "../../../../../e2eTestUtils/LabApiQueryParams";
 import { AzureEnvironments, AppTypes } from "../../../../../e2eTestUtils/Constants";
 import { LabClient } from "../../../../../e2eTestUtils/LabClient";
-import { msalConfig as aadMsalConfig, request as aadTokenRequest } from "../authConfig.json";
-import { getBrowser, getHomeUrl, clickLoginPopup, clickLoginRedirect, waitForReturnToApp } from "../../testUtils";
+import { msalConfig as aadMsalConfig, request as aadTokenRequest } from "../authConfigs/localStorageAuthConfig.json";
+import { clickLoginPopup, clickLoginRedirect, waitForReturnToApp } from "./testUtils";
 import fs from "fs";
+import {getBrowser, getHomeUrl} from "../../testUtils";
 
-const SCREENSHOT_BASE_FOLDER_NAME = `${__dirname}/screenshots`;
-let sampleHomeUrl = "";
+const SCREENSHOT_BASE_FOLDER_NAME = `${__dirname}/screenshots/localStorageTests`;
 
 async function verifyTokenStore(BrowserCache: BrowserCacheUtils, scopes: string[]): Promise<void> {
     const tokenStore = await BrowserCache.getTokens();
@@ -25,6 +25,7 @@ async function verifyTokenStore(BrowserCache: BrowserCacheUtils, scopes: string[
 describe("LocalStorage Tests", function () {
     let username = "";
     let accountPwd = "";
+    let sampleHomeUrl = "";
 
     let browser: puppeteer.Browser;
     beforeAll(async () => {
@@ -42,7 +43,7 @@ describe("LocalStorage Tests", function () {
 
         [username, accountPwd] = await setupCredentials(envResponse[0], labClient);
 
-        fs.writeFileSync("./app/customizable-e2e-local-storage/testConfig.json", JSON.stringify({msalConfig: aadMsalConfig, request: aadTokenRequest}));
+        fs.writeFileSync("./app/customizable-e2e-test/testConfig.json", JSON.stringify({msalConfig: aadMsalConfig, request: aadTokenRequest}));
     });
 
     let context: puppeteer.BrowserContext;
