+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/multipleResources/jest.config.js
@@ -1,8 +1,8 @@
 module.exports = {
     displayName: "Vanilla JS multiple resources",
     globals: {
-        __PORT__: 30665,
-        __STARTCMD__: "npm start -- --port 30665 --sample multipleResources",
+        __PORT__: 30662,
+        __STARTCMD__: "npm start -- --port 30662 --sample multipleResources",
     },
     testMatch: [
         "<rootDir>/test/**/*.spec.ts",
