+++ samples/msal-browser-samples/VanillaJSTestApp2.0/package.json
@@ -8,7 +8,7 @@
     "start": "node server.js",
     "build:package": "cd ../../../lib/msal-browser && npm run build:all",
     "start:build": "npm run build:package && npm start",
-    "test:e2e": "jest",
+    "test:e2e": "jest --runInBand",
     "generate:certs": "openssl.exe req -x509 -newkey rsa:4096 -keyout key.pem -out cert.pem -days 7 -nodes -subj /CN=localhost"
   },
   "dependencies": {
